refactor out integration tests
