+++ bn254/src/compression.rs
@@ -278,7 +278,6 @@ mod tests {
         super::*,
         crate::compression::target_arch::convert_endianness,
         ark_serialize::{CanonicalDeserialize, CanonicalSerialize, Compress, Validate},
-        serde_derive::Deserialize,
         std::ops::Neg,
         target_arch::{
             alt_bn128_g1_compress, alt_bn128_g1_decompress, alt_bn128_g2_compress,
@@ -387,30 +386,5 @@ mod tests {
         let g1_compressed = alt_bn128_g2_compress(&g1_bytes).unwrap();
         let g1_decompressed = alt_bn128_g2_decompress(&g1_compressed).unwrap();
         assert_eq!(g1_bytes, g1_decompressed);
-    }
-    #[test]
-    fn alt_bn128_compression_pairing_test_input() {
-        let test_data = include_str!("../tests/data/pairing_cases.json");
-
-        #[derive(Deserialize)]
-        #[serde(rename_all = "PascalCase")]
-        struct TestCase {
-            input: String,
-        }
-
-        let test_cases: Vec<TestCase> = serde_json::from_str(test_data).unwrap();
-
-        test_cases.iter().for_each(|test| {
-            let input = array_bytes::hex2bytes_unchecked(&test.input);
-            if input.len() < 192 {
-                return;
-            }
-            let g1 = input[0..64].to_vec();
-            let g1_compressed = alt_bn128_g1_compress(&g1).unwrap();
-            assert_eq!(g1, alt_bn128_g1_decompress(&g1_compressed).unwrap());
-            let g2 = input[64..192].to_vec();
-            let g2_compressed = alt_bn128_g2_compress(&g2).unwrap();
-            assert_eq!(g2, alt_bn128_g2_decompress(&g2_compressed).unwrap());
-        });
     }
 }
