+++ bn254/src/lib.rs
@@ -447,7 +447,6 @@ mod tests {
         ark_bn254::g1::G1Affine,
         ark_ec::AffineRepr,
         ark_serialize::{CanonicalSerialize, Compress},
-        serde_derive::Deserialize,
     };
 
     #[test]
@@ -468,74 +467,6 @@ mod tests {
             .try_into()
             .unwrap();
         assert_eq!(p, zero);
-    }
-
-    #[test]
-    fn alt_bn128_addition_test() {
-        let test_data = include_str!("../tests/data/addition_cases.json");
-
-        #[derive(Deserialize)]
-        #[serde(rename_all = "PascalCase")]
-        struct TestCase {
-            input: String,
-            expected: String,
-        }
-
-        let test_cases: Vec<TestCase> = serde_json::from_str(test_data).unwrap();
-
-        test_cases.iter().for_each(|test| {
-            let input = array_bytes::hex2bytes_unchecked(&test.input);
-            let result = alt_bn128_addition(&input);
-            assert!(result.is_ok());
-
-            let expected = array_bytes::hex2bytes_unchecked(&test.expected);
-
-            assert_eq!(result.unwrap(), expected);
-        });
-    }
-
-    #[test]
-    fn alt_bn128_multiplication_test() {
-        let test_data = include_str!("../tests/data/multiplication_cases.json");
-
-        #[derive(Deserialize)]
-        #[serde(rename_all = "PascalCase")]
-        struct TestCase {
-            input: String,
-            expected: String,
-        }
-
-        let test_cases: Vec<TestCase> = serde_json::from_str(test_data).unwrap();
-
-        test_cases.iter().for_each(|test| {
-            let input = array_bytes::hex2bytes_unchecked(&test.input);
-            let result = alt_bn128_multiplication(&input);
-            assert!(result.is_ok());
-            let expected = array_bytes::hex2bytes_unchecked(&test.expected);
-            assert_eq!(result.unwrap(), expected);
-        });
-    }
-
-    #[test]
-    fn alt_bn128_pairing_test() {
-        let test_data = include_str!("../tests/data/pairing_cases.json");
-
-        #[derive(Deserialize)]
-        #[serde(rename_all = "PascalCase")]
-        struct TestCase {
-            input: String,
-            expected: String,
-        }
-
-        let test_cases: Vec<TestCase> = serde_json::from_str(test_data).unwrap();
-
-        test_cases.iter().for_each(|test| {
-            let input = array_bytes::hex2bytes_unchecked(&test.input);
-            let result = alt_bn128_pairing(&input);
-            assert!(result.is_ok());
-            let expected = array_bytes::hex2bytes_unchecked(&test.expected);
-            assert_eq!(result.unwrap(), expected);
-        });
     }
 
     #[test]
